Remove observable models and squash observability into chat models
