+++ langchain4j/src/test/java/dev/langchain4j/service/AiServicesIT.java
@@ -4,7 +4,6 @@ import dev.langchain4j.exception.IllegalConfigurationException;
 import dev.langchain4j.memory.ChatMemory;
 import dev.langchain4j.memory.chat.MessageWindowChatMemory;
 import dev.langchain4j.model.chat.ChatLanguageModel;
-import dev.langchain4j.model.chat.listener.ObservableChatModel;
 import dev.langchain4j.model.chat.request.ChatRequest;
 import dev.langchain4j.model.input.structured.StructuredPrompt;
 import dev.langchain4j.model.moderation.ModerationModel;
@@ -89,12 +88,10 @@ class AiServicesIT {
         } catch (Throwable ignored) {
             // don't care if it was called or not
         }
-        if (model instanceof ObservableChatModel observableChatModel) {
-            try {
-                verify(observableChatModel).listeners();
-            } catch (Throwable ignored) {
-                // don't care if it was called or not
-            }
+        try {
+            verify(model).listeners();
+        } catch (Throwable ignored) {
+            // don't care if it was called or not
         }
         verifyNoMoreInteractions(model);
     }
