+++ langchain4j/src/test/java/dev/langchain4j/service/StreamingAiServicesWithToolsIT.java
@@ -11,7 +11,6 @@ import dev.langchain4j.data.message.AiMessage;
 import dev.langchain4j.data.message.ChatMessage;
 import dev.langchain4j.memory.ChatMemory;
 import dev.langchain4j.memory.chat.MessageWindowChatMemory;
-import dev.langchain4j.model.chat.listener.ObservableStreamingChatModel;
 import dev.langchain4j.model.chat.StreamingChatLanguageModel;
 import dev.langchain4j.model.chat.request.ChatRequest;
 import dev.langchain4j.model.chat.request.json.JsonObjectSchema;
@@ -373,12 +372,10 @@ class StreamingAiServicesWithToolsIT {
         } catch (Throwable ignored) {
             // don't care if it was called or not
         }
-        if (model instanceof ObservableStreamingChatModel observableStreamingChatModel) {
-            try {
-                verify(observableStreamingChatModel).listeners();
-            } catch (Throwable ignored) {
-                // don't care if it was called or not
-            }
+        try {
+            verify(model).listeners();
+        } catch (Throwable ignored) {
+            // don't care if it was called or not
         }
         verifyNoMoreInteractions(model);
     }
